refactor(scheduler): isolate database package
